LayoutTests/js/typedarray-indexOf-expected.txt

This test checks the behavior of the TypedArray.prototype.indexOf function

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Int32Array.prototype.indexOf.length is 1
PASS Int32Array.prototype.indexOf.name is 'indexOf'
PASS isSameFunctionForEachTypedArrayPrototype('indexOf') is true
PASS testPrototypeReceivesArray('indexOf', [undefined, this, { }, [ ], true, '']) is true

testPrototypeFunction has the following arg list (name, args, init, result [ , expectedArray ])

PASS testPrototypeFunction('indexOf', '(2, -500)', array, 0) is true
PASS testPrototypeFunction('indexOf', '(9, 500)', array, -1) is true
PASS testPrototypeFunction('indexOf', '(2)', array, 0) is true
PASS testPrototypeFunction('indexOf', '(7)', array, -1) is true
PASS testPrototypeFunction('indexOf', '(2, 3)', array, 3) is true
PASS testPrototypeFunction('indexOf', '(2, 2)', array, 3) is true
PASS testPrototypeFunction('indexOf', '(2, 0)', array, 0) is true
PASS testPrototypeFunction('indexOf', '(2, -1)', array, 3) is true
PASS testPrototypeFunction('indexOf', '(2, -2)', array, 3) is true

PASS successfullyParsed is true

TEST COMPLETE
